fix: integrate staking status filter to combined
